+++ controlplane/src/core/repositories/SchemaCheckRepository.ts
@@ -22,6 +22,9 @@ export class SchemaCheckRepository {
     isComposable?: boolean;
     isDeleted?: boolean;
     proposedSubgraphSchemaSDL: string;
+    trafficCheckSkipped?: boolean;
+    lintSkipped?: boolean;
+    graphPruningSkipped?: boolean;
   }): Promise<string> {
     const insertedSchemaCheck = await this.db
       .insert(schemaChecks)
@@ -30,6 +33,9 @@ export class SchemaCheckRepository {
         isComposable: data.isComposable,
         isDeleted: data.isDeleted,
         proposedSubgraphSchemaSDL: data.proposedSubgraphSchemaSDL,
+        clientTrafficCheckSkipped: data.trafficCheckSkipped || false,
+        lintSkipped: data.lintSkipped || false,
+        graphPruningSkipped: data.graphPruningSkipped || false,
       })
       .returning()
       .execute();