+++ studio/src/components/checks/lint-issues-table.tsx
@@ -37,18 +37,17 @@ export const LintIssuesTable = ({
     return (
       <EmptyState
         icon={<NoSymbolIcon className="text-gray-400" />}
-        title="Schema Linting is disabled"
-        description="Enable schema linting to catch linting issues in your schema."
+        title="Schema Linting Skipped"
+        description="Linting was skipped for this run. Configure it to catch linting issues in your schema."
         actions={
           <Button
-            size="sm"
             onClick={() => {
               router.push(
                 `/${router.query.organizationSlug}/lint-policy?namespace=${router.query.namespace}`,
               );
             }}
           >
-            Navigate to Lint Policy
+            Configure Lint Policy
           </Button>
         }
       />