+++ cli/CHANGELOG.md
@@ -4,6 +4,16 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.66.1](https://github.com/wundergraph/cosmo/compare/wgc@0.66.0...wgc@0.66.1) (2024-09-25)
+
+**Note:** Version bump only for package wgc
+
+# [0.66.0](https://github.com/wundergraph/cosmo/compare/wgc@0.65.1...wgc@0.66.0) (2024-09-23)
+
+### Features
+
+* option to skip traffic check in schema check ([#1211](https://github.com/wundergraph/cosmo/issues/1211)) ([d9fd83a](https://github.com/wundergraph/cosmo/commit/d9fd83ab99d2f2fc8e5f99a4cdb3abb8a1a7837c)) (@thisisnithin)
+
 ## [0.65.1](https://github.com/wundergraph/cosmo/compare/wgc@0.65.0...wgc@0.65.1) (2024-09-20)
 
 **Note:** Version bump only for package wgc